+++ RedProtect-Spigot/src/main/java/br/net/fabiozumbi12/RedProtect/Bukkit/helpers/EconomyManager.java
@@ -1,6 +1,6 @@
 /*
- * Copyright (c) 2012-2023 - @FabioZumbi12
- * Last Modified: 02/10/2023 22:14
+ * Copyright (c) 2012-2024 - @FabioZumbi12
+ * Last Modified: 26/06/2024 16:59
  *
  * This class is provided 'as-is', without any express or implied warranty. In no event will the authors be held liable for any
  *  damages arising from the use of this class.
@@ -88,11 +88,11 @@ public class EconomyManager {
         return RedProtect.get().getLanguageManager().get("economy.forsale") + " &6" + getFormatted(r.getValue()) + " &2" + RedProtect.get().getConfigManager().ecoRoot().economy_name;
     }
 
-    public static String getFormatted(long value) {
+    public static String getFormatted(double value) {
         return RedProtect.get().getConfigManager().ecoRoot().economy_symbol + value;
     }
 
-    public static boolean putToSell(Region r, String uuid, long value) {
+    public static boolean putToSell(Region r, String uuid, double value) {
         try {
             r.clearMembers();
             r.clearAdmins();
